Changes for creating platform specific jars and a single fat-framework file

+++ src/iosMain/kotlin/com/twelfthmile/kyuga/expectations/getMaxDateYear.kt
@@ -11,5 +11,5 @@ actual fun getMaxDateYear(): Long {
     val calender = NSCalendar.currentCalendar
     val unitFlags = NSYearCalendarUnit or NSMonthCalendarUnit or NSDayCalendarUnit
     val components = calender.components(unitFlags, currentDate)
-    return components.year
+    return components.year.toLong()
 }
